First version.
This version detects objects with YOLOv8 and uses the DeepSort algorithm to track objects in the video. Currently, the detection is quite accurate and it can identify objects that were lost from tracking for up to 350 frames.

It has a functionality to save data in Excel, but it does not yet assign the end time of the bin’s presence